Improve navigation visibility when scrolling
When scrolling, the navigation header was fixed to the top of the
viewport. Due to how it's positioned, it goes behind some images.

Additionally, the background color was set to transparent. This means
that the navigation text becomes unreadable when it's on top of other
text.

By changing the z-index, setting background-color to white, and adding
some padding, the navigation header becomes a bit more usable.